Alan found a deadlock in TAU when -MULTIPLECOUNTERS -papi=<dir> -TRACE was used. This problem was narrowed down to the initialization routine where we use RtsLayer::LockDB() and then initialize a bunch of data. When tracing is used with multiplecounters, in addition, we initialize an array of user defined events that track COUNTER2-N with each routine transition. The problem is, inside the constructor of TauUserEvent, we invoke TauUserEvent::AddEventToDB() which calls LockDB() again. This problem stems from the fact that we have only one lock and we use it for locking FunctionInfo, MultiCounterLayer and TauUserEvents! So, in building these objects, we found a case where if the user defined events had its own lock, it would not interfere with the other locks. If (or rather When) we redesign TAU's measurement substrate, we must implement multiple locks so that this does not occur. The reason we didn't have multiple locks was because each thread package had its own datatype for locks and we'd have had to deal with converting to a void * or finding a generic way to handle multiple locks.
For now, we release the lock then call a bunch of user defined event ctors (corresponding to COUNTER2-N) and then re-acquire it before leaving the thread section. This is a sub-optimal solution and one that merits further investigation.
I need to ask for Robert's advice on how best to restructure the locking code
in multiple counters.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@2674 a8e31b63-e6ca-409b-a01b-2eec1b506579

+++ src/Profile/MultipleCounters.cpp
@@ -899,7 +899,7 @@ void MultipleCounterLayer::triggerCounterEvents(unsigned long long timestamp, do
   int i;
   for (i = 1; i < numberOfActiveFunctions; i++)
   { /* for each event */
-    TraceEvent(counterEvents[i]->GetEventId(), values[i], tid, timestamp, 1);
+    TraceEvent(counterEvents[i]->GetEventId(), (long long) values[i], tid, timestamp, 1);
     // 1 in the last parameter is for use timestamp 
   }
